Fix weird wording
The literal English translation to that line would have been "The code
next".

Now, in Spanish adjectives can go either before or after the noun, but
you'd be hard pressed to find anyone saying "siguiente" here after the
noun since it just doesn't sound right. With this change it's "The next
code" instead.